Changes to fit with game style
Added the "isWeapon" attribute to the Item Class so that we can create weapons for the players to use during combat.
Add the "Damage" attribute to items as well to determine how much damage the weapons will do when combat is implemented
- Added minimum and maximum damage so a random number can be generated within the damage range

Altered the game to make sense in a dungeon related context (rooms and Items).